Fix portability issue with sgen garbage collector
While the Bohem garbage collector works with V8.Net sgen has an issue
with the statement InternalHandle.cs:111 and InternalHandle.cs:119 in
order to fix this we needed to store InPtr's instead of HandleProxy*
pointers.

Also we needed change the test to make them pass on a x64 bit system.